feat(useApi): type parameters & thorough docs
- type the fetcher
- type the errors
- pass the type parameters to useSWR

Note that the response type is still asserted,
no actual parsing happens
